Fixed movement working based on the rotation of the MainCharacter instead of the Camera
In the SpringArm settings I unticked inheriting from X, Y, Z from the parent component, as I didn't want the camera following where the MainCharacter where it's looking. This resulted in the movement being messed up as it was still basing it off the direction the MainCharacter was facing.
So, I changed it so in the Control Rotation function, it gets world rotation based off the camera.