Fire functionality works
I created an arrow on the BP_MainCharacter, that is used as reference point as to where the BP_ProjectileBase is spawned from for the Firing Functionality. Then, in Blueprint I got the reference for the arrow called 'ShootDirection', got its location and Rotation, added them together, converted them to a Transform value, then plugged it in the Spawn Actor that will spawn the Projectile when Input Fire is active.