Made it so a projectile destroys itself after a few secs if it doesn't hit anything, then in the Enemy class I made it so
In the BP_Enemy class, I have added on top of the EventDamage check to destroy the actor, a SpawnActor from class so I can spawn a collectible in the place of the destroyed enemy. I had to change the way I have done, as I tried to get the world location after the enemy actor was already destoyed, so what I did instead, I se the collision and visiblity to disabled, then got the world location, Spawned the collactable in, then destroyed the BP_Enemy actor.